Rework storage API connection
Changes:
- Copy gRPC connection state tracking used for VAL API to the
  storage client and radio API classes.
- Use the connection state in the storage client to update a new
  RiverPod storageClientConnected notifier.
- Use the notifier in the users and units code to prevent blocking
  on storage API operations if the API backend is not connected.
- Unified the naming of the various API backend providers' async
  connection methods to "connect", and reworked things so they are
  all called in main.  The profile-driven user and units
  initialization has been moved to happen when the storage API
  connects.  This seems to have the benefit of making the splash
  screen behavior more consistent.
- Rework initialization in the Users provider to default to
  a no users state, and only initialize with the default demo
  users if the storage API is available and empty.
- Use the storageiClientConnected notifier to gate whether the
  profiles option is visible on the settings page.  Implementing
  some form of fallback profile storage scheme when the storage
  API is unavailable has been deemed not worth implementing.
- Switched to using watch instead of read for the voice agent
  enabled state in the settings page code.  In practice this
  likely does not make much difference as the configuration file
  reading will happen early enough from other usage to not race,
  but watch is more correct in widget context.
- Cleaned up a misspelling in some variable names.

Bug-AGL: SPEC-5259

+++ lib/data/data_providers/users_notifier.dart
@@ -5,7 +5,6 @@ import 'package:uuid/uuid.dart';
 import '../models/user.dart';
 
 import 'package:protos/storage-api.dart' as storage_api;
-import 'initialize_settings.dart';
 
 class UsersNotifier extends Notifier<Users> {
   @override
@@ -18,7 +17,6 @@ class UsersNotifier extends Notifier<Users> {
   Users build() {
     // Initialize default state.
     state = Users.initial();
-    loadUsers();
     return state;
   }
 
@@ -31,25 +29,15 @@ class UsersNotifier extends Notifier<Users> {
       if (searchResponseUsers.result.isEmpty) {
         // Add default users if no users are inside the storage API.
         debugPrint("Adding default demo user profiles");
-        loadUsers();
-        await storageClient.write(storage_api.KeyValue(
-            key: '${UsersPath.InfotainmentUsers}.${_users[0].id}.id',
-            value: _users[0].id));
-        await storageClient.write(storage_api.KeyValue(
-            key: '${UsersPath.InfotainmentUsers}.${_users[0].id}.name',
-            value: _users[0].name));
-        await storageClient.write(storage_api.KeyValue(
-            key: '${UsersPath.InfotainmentUsers}.${_users[1].id}.id',
-            value: _users[1].id));
-        await storageClient.write(storage_api.KeyValue(
-            key: '${UsersPath.InfotainmentUsers}.${_users[1].id}.name',
-            value: _users[1].name));
-        await storageClient.write(storage_api.KeyValue(
-            key: '${UsersPath.InfotainmentUsers}.${_users[2].id}.id',
-            value: _users[2].id));
-        await storageClient.write(storage_api.KeyValue(
-            key: '${UsersPath.InfotainmentUsers}.${_users[2].id}.name',
-            value: _users[2].name));
+        state = state.copyWith(users: _users);
+        for (int i = 0; i < _users.length; i++ ) {
+          await storageClient.write(storage_api.KeyValue(
+              key: '${UsersPath.InfotainmentUsers}.${_users[i].id}.id',
+              value: _users[i].id));
+          await storageClient.write(storage_api.KeyValue(
+              key: '${UsersPath.InfotainmentUsers}.${_users[i].id}.name',
+              value: _users[i].name));
+        }
         await selectUser(_users[0].id);
       } else {
         List<User> users = [];
@@ -84,33 +72,29 @@ class UsersNotifier extends Notifier<Users> {
     } catch (e) {
       // Fallback to initial defaults if error.
       debugPrint('Error loading users: $e');
-      loadUsers();
-      state = state.copyWith(selectedUser: _users[0]);
     }
-  }
-
-  void loadUsers() {
-    state = state.copyWith(users: _users);
   }
 
   Future<void> selectUser(String userId) async {
-    final storageClient = ref.read(storageClientProvider);
     var selectedUser = state.users.firstWhere((user) => user.id == userId);
     state = state.copyWith(selectedUser: selectedUser);
-    // Write to storage API.
-    try {
-      await storageClient.write(storage_api.KeyValue(
-        key: UsersPath.InfotainmentCurrentUser,
-        value: userId,
-      ));
-    } catch (e) {
-      debugPrint('Error selecting user: $e');
-    }
 
-    try {
-      await initializeSettingsUser(ref);
-    } catch (e) {
-      debugPrint('Error loading settings of user: $e');
+    // Write to storage API.
+    if (ref.read(storageClientConnectedProvider)) {
+      try {
+        await ref.read(storageClientProvider).write(storage_api.KeyValue(
+          key: UsersPath.InfotainmentCurrentUser,
+          value: userId,
+        ));
+      } catch (e) {
+        debugPrint('Error selecting user: $e');
+      }
+
+      try {
+        await ref.read(unitStateProvider.notifier).loadSettingsUnits();
+      } catch (e) {
+        debugPrint('Error loading settings of user: $e');
+      }
     }
   }
 
@@ -122,47 +106,55 @@ class UsersNotifier extends Notifier<Users> {
     if (state.users.isNotEmpty && currentUserId == userId) {
       state = state.copyWith(selectedUser: state.users.first);
       // Write to API to change selected user.
-      await storageClient.write(storage_api.KeyValue(
-          key: UsersPath.InfotainmentCurrentUser, value: state.users.first.id));
+      if (ref.read(storageClientConnectedProvider)) {
+        await storageClient.write(storage_api.KeyValue(
+            key: UsersPath.InfotainmentCurrentUser, value: state.users.first.id));
+      }
     }
     if (state.users.isEmpty) {
       state = state.copyWith(selectedUser: const User(id: '0', name: ''));
       // Write to API to change selected user.
-      await storageClient.write(storage_api.KeyValue(
-          key: UsersPath.InfotainmentCurrentUser, value: '0'));
+      if (ref.read(storageClientConnectedProvider)) {
+        await storageClient.write(storage_api.KeyValue(
+            key: UsersPath.InfotainmentCurrentUser, value: '0'));
+      }
     }
     // Delete from storage API.
-    try {
-      final searchResponse =
-          await storageClient.search(storage_api.Key(key: userId));
-      final keyList = searchResponse.result;
-      // Delete id, name entries of the user from the default namespace.
-      for (final key in keyList) {
-        await storageClient.delete(storage_api.Key(key: key));
-      }
-      // Delete all VSS keys from the user namespace.
-      await storageClient
+    if (ref.read(storageClientConnectedProvider)) {
+      try {
+        final searchResponse =
+            await storageClient.search(storage_api.Key(key: userId));
+        final keyList = searchResponse.result;
+        // Delete id, name entries of the user from the default namespace.
+        for (final key in keyList) {
+          await storageClient.delete(storage_api.Key(key: key));
+        }
+        // Delete all VSS keys from the user namespace.
+        await storageClient
           .deleteNodes(storage_api.Key(key: "Vehicle", namespace: userId));
-    } catch (e) {
-      debugPrint('Error removing user with id $userId: $e');
+      } catch (e) {
+        debugPrint('Error removing user with id $userId: $e');
+      }
     }
   }
 
   Future<void> addUser(String userName) async {
-    final storageClient = ref.read(storageClientProvider);
     final id = const Uuid().v1();
     final user = User(id: id, name: userName);
     state.users.insert(0, user);
     // New user is automatically selected.
     await selectUser(user.id);
     // Write to storage API.
-    try {
-      await storageClient.write(storage_api.KeyValue(
-          key: '${UsersPath.InfotainmentUsers}.$id.name', value: userName));
-      await storageClient.write(storage_api.KeyValue(
-          key: '${UsersPath.InfotainmentUsers}.$id.id', value: id));
-    } catch (e) {
-      debugPrint('Error adding user with id $id: $e');
+    if (ref.read(storageClientConnectedProvider)) {
+      final storageClient = ref.read(storageClientProvider);
+      try {
+        await storageClient.write(storage_api.KeyValue(
+            key: '${UsersPath.InfotainmentUsers}.$id.name', value: userName));
+        await storageClient.write(storage_api.KeyValue(
+            key: '${UsersPath.InfotainmentUsers}.$id.id', value: id));
+      } catch (e) {
+        debugPrint('Error adding user with id $id: $e');
+      }
     }
   }
